+++ pages/edit/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1" />
-    <title>JavaScript 2 CA | Register New User</title>
+    <title>JavaScript 2 CA | Edit Post</title>
     <link rel="stylesheet" href="/dist/css/style.css" />
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.1/dist/js/bootstrap.bundle.min.js" defer integrity="sha384-u1OknCvxWvY5kfmNBILK2hRnQC3Pr17a+RTT6rIHI7NnikvbZlHgTPOOmMi466C8" crossorigin="anonymous"></script>
     <script defer type="module" src="/src/js/index.mjs"></script>
@@ -43,7 +43,7 @@
         </div>
       </nav>
     </header>
-    <div class="container row m-auto h-100 d-flex">
+    <div class="container row m-auto h-75 d-flex">
       <main class="col-10 col-sm-7 col-md-5 m-auto text-center d-flex align-items-center">
         <form id="editPost" class="flex-grow-1">
           <h1 class="h3 mb-5 fw-normal text-primary text-uppercase fw-bold m-auto">Edit</h1>
@@ -51,14 +51,16 @@
             <div class="form-group">
               <textarea required name="body" class="form-control social-media-post bg-secondary text-primary" style="--bs-bg-opacity: 0.4" id="message" rows="3" placeholder="Hey, what's up?"></textarea>
             </div>
-            <div class="form-group mt-3">
+            <!-- Took these out because of bad request: "body/media must match format"-->
+            <!-- <div class="form-group mt-3">
               <label for="" class="text-primary">Tags</label>
               <input type="text" name="tags" class="form-control" />
             </div>
+
             <div class="form-group mt-3">
               <label for="" class="text-primary">Image</label>
               <input type="url" name="media" class="form-control" />
-            </div>
+            </div> -->
           </div>
           <button class="ps-5 pe-5 mt-5 btn btn-lg btn-primary text-uppercase" type="submit">Update</button>
         </form>
