+++ src/js/api/posts/create.mjs
@@ -1,7 +1,5 @@
 import { API_SOCIAL_URL } from "../constants.mjs";
 import { authFetch } from "../authFetch.mjs";
-import { getHomeFeedPosts } from "../../handlers/getPosts.mjs";
-import { getProfileFeedPosts } from "../../handlers/getPosts.mjs";
 
 const action = "/posts";
 
@@ -16,27 +14,8 @@ export async function createPost(postData) {
 
     const post = await response.json();
     console.log(post);
+    return post;
   } catch (error) {
     console.log(error);
   }
 }
-
-const button = document.querySelector(".submit-post");
-const input = document.querySelector("textarea");
-
-button.addEventListener("click", (e) => {
-  e.preventDefault();
-
-  const bodyText = {
-    title: "No title",
-    body: input.value,
-  };
-  createPost(bodyText);
-  input.value = "";
-
-  if (location.pathname === "/pages/profile/") {
-    getProfileFeedPosts();
-  } else {
-    getHomeFeedPosts();
-  }
-});
