+++ src/js/index.mjs
@@ -1,7 +1,8 @@
 import { setRegisterFormListener } from "./handlers/register.mjs";
 import { setLoginFormListener } from "./handlers/login.mjs";
-import { getHomeFeedPosts } from "./handlers/getPosts.mjs";
-import { getProfileFeedPosts } from "./handlers/getPosts.mjs";
+import { setUpdatePostListener } from "./handlers/updatePost.mjs";
+import { setCreatePostListener } from "./handlers/createPost.mjs";
+import { getHomeFeedPosts, getProfileFeedPosts } from "./handlers/getPosts.mjs";
 
 const path = location.pathname;
 
@@ -10,9 +11,11 @@ if (path === "/pages/register/") {
 } else if (path === "/pages/login/") {
   setLoginFormListener();
 } else if (path === "/pages/profile/") {
+  setCreatePostListener();
   getProfileFeedPosts();
 } else if (path === "/pages/edit/") {
-  console.log("edit");
+  setUpdatePostListener();
 } else {
-  //   getHomeFeedPosts();
+  setCreatePostListener();
+  getHomeFeedPosts();
 }
